+++ src/components/Faq.tsx
@@ -1,8 +1,8 @@
 import { QAData } from '@/lib/FaqData';
 import styles from '@/styles/components/Faq.module.css';
 import { useState } from 'react';
-import Accordian from '@/components/FaqAccordian';
-import Secret from '@/components/secret';
+import Accordian from '@/components/FaqAccordiansad';
+import FaqAccordian from '@/components/FaqAccordian';
 
 const Faq = () => {
 
@@ -10,62 +10,16 @@ const Faq = () => {
     let QAData1 = QAData.slice(0, halfIndex); //force two columns
     let QAData2 = QAData.slice(halfIndex);
 
-    const [active1, setActive1] = useState<boolean[]>([false, false, false, false]);
-    const [active2, setActive2] = useState<boolean[]>([false, false, false, false])
+    const [active, setActive] = useState<boolean[]>([false, false, false, false, false, false, false, false]);
 
 	return (
-        <div className={styles['faq-content']}>
-            <div className={styles['faq-div']}>
-                {QAData1.map((el, i) => {
-                    return (
-                        <div key={"questions" + i}>
-                            <Secret
-                            question={el.question}
-                            answer={el.answer}
-                            turn={active1}
-                            setTurn={setActive1}
-                            idx={el.idx}
-                            />
-                        </div>
-                    )
-                })}
-            </div>
-            <div className={styles['faq-div']}>
-                {QAData2.map((el, i) => {
-                    return (
-                        <div key={"questions" + i}>
-                            <Secret
-                            question={el.question}
-                            answer={el.answer}
-                            turn={active2}
-                            setTurn={setActive2}
-                            idx={el.idx}
-                            />
-                        </div>
-                    )
-                })}
-            </div>
-            {/* <div className={styles.faqcontent}>
-            <div className={styles.column}>
-                {QAData1.map((el, i) => {
+        <div className={styles['faq-background']}>
+            <div className={styles['faq-content']}>
+                <div className={styles['faq-div']}>
+                    {QAData1.map((el, i) => {
                         return (
-                            <div className='w-full' key={"questions" + i}>
-                                <Accordian
-                                question={el.question}
-                                answer={el.answer}
-                                turn={active}
-                                setTurn={setActive}
-                                idx={el.idx}
-                                />
-                            </div>
-                        )
-                })}
-            </div>
-            <div className={styles.column}>
-                {QAData2.map((el, i) => {
-                        return (
-                            <div className='w-full' key={"questions" + i + halfIndex}>
-                                <Accordian
+                            <div key={"questions" + i}>
+                                <FaqAccordian
                                 question={el.question}
                                 answer={el.answer}
                                 turn={active}
@@ -75,8 +29,23 @@ const Faq = () => {
                             </div>
                         )
                     })}
+                </div>
+                <div className={styles['faq-div']}>
+                    {QAData2.map((el, i) => {
+                        return (
+                            <div key={"questions" + (i+halfIndex)}>
+                                <FaqAccordian
+                                question={el.question}
+                                answer={el.answer}
+                                turn={active}
+                                setTurn={setActive}
+                                idx={el.idx}
+                                />
+                            </div>
+                        )
+                    })}
+                </div>
             </div>
-        </div> */}
         </div>
 	);
 };
