+++ src/components/FaqAccordian.tsx
@@ -10,21 +10,10 @@ type Props = {
     idx: number
 }
 
-const Accordian = ({question, answer, turn, setTurn, idx}: Props) => {
-
-    const contentRef = useRef<HTMLDivElement>(null);
-
-    useEffect(()=>{
-        if(contentRef.current) {
-            //contentRef.current.style.maxHeight = turn![idx] ? `${contentRef.current.scrollHeight}px` : "0px";
-            contentRef.current.style.opacity = turn![idx] ? '100%' : '0%';
-            contentRef.current.style.fontSize = turn![idx] ? '18px' : '0px';
-        }
-    }, [contentRef, turn, idx])
+const FaqAccordian = ({question, answer, turn, setTurn, idx}: Props) => {
 
     const toggleAccordian = () => { // close all other
         let newTurn = [...turn!];
-        //newTurn[idx] = !newTurn[idx]
         let turnIdx = !newTurn[idx];
         newTurn = newTurn.fill(false);
         newTurn[idx] = turnIdx;
@@ -32,7 +21,27 @@ const Accordian = ({question, answer, turn, setTurn, idx}: Props) => {
     }
 
     return (
-        <div className={styles.accordianContent}>
+        <div>
+            <button onClick={toggleAccordian}
+            className={styles['faq-button']}
+            >
+            <div className={styles['faq-bullet']}>
+                { turn![idx] ?
+                <img className={styles['bullet-img']} src={bulletExpand.src} alt=""/> :
+                <img className={styles['bullet-img']} src={bullet.src} alt=""/>
+                }
+                <p className={styles['faq-question']}>{question}</p>
+            </div>
+            </button>
+            <div className={styles['faq-answer']}
+            style={
+                turn![idx]
+                ? { maxHeight: "60px" } :
+                { maxHeight: "0"}
+            }>
+                <p>{answer}</p>
+            </div>
+        {/* <div className={styles.accordianContent}>
             <button onClick={toggleAccordian}
                     className={styles.accordianButton}>
                     <div>
@@ -49,8 +58,9 @@ const Accordian = ({question, answer, turn, setTurn, idx}: Props) => {
                         </p>
                     </div>
             </button>
+        </div> */}
         </div>
     )
 }
 
-export default Accordian;
+export default FaqAccordian;
